moved funnel queries to a new folder for better organization; separated funnel_persons and funnel_trends_persons into individual classes;
