+++ ee/clickhouse/queries/__init__.py
@@ -1,4 +1,5 @@
-from .clickhouse_funnel import ClickhouseFunnel
+from ee.clickhouse.queries.funnels.funnel import ClickhouseFunnel
+
 from .clickhouse_paths import ClickhousePaths
 from .clickhouse_retention import ClickhouseRetention
 from .clickhouse_session_recording import SessionRecording
