--- ee/clickhouse/queries/clickhouse_funnel_base.py
@@ -1,44 +1,0 @@
-from ee.clickhouse.models.action import format_action_filter
-from ee.clickhouse.models.property import parse_prop_clauses
-from posthog.constants import TREND_FILTER_TYPE_ACTIONS
-from posthog.models import Action, Entity, Filter, Team
-from posthog.queries.funnel import Funnel
-
-
-class ClickhouseFunnelBase(Funnel):
-    _filter: Filter
-    _team: Team
-
-    def __init__(self, filter: Filter, team: Team) -> None:
-        self._filter = filter
-        self._team = team
-        self.params = {
-            "team_id": self._team.pk,
-            "events": [],  # purely a speed optimization, don't need this for filtering
-        }
-
-    def _build_steps_query(self, entity: Entity, index: int) -> str:
-        filters = self._build_filters(entity, index)
-        if entity.type == TREND_FILTER_TYPE_ACTIONS:
-            action = Action.objects.get(pk=entity.id)
-            for action_step in action.steps.all():
-                self.params["events"].append(action_step.event)
-            action_query, action_params = format_action_filter(action, "step_{}".format(index))
-            if action_query == "":
-                return ""
-
-            self.params.update(action_params)
-            content_sql = "{actions_query} {filters}".format(actions_query=action_query, filters=filters,)
-        else:
-            self.params["events"].append(entity.id)
-            content_sql = "event = '{event}' {filters}".format(event=entity.id, filters=filters)
-        return content_sql
-
-    def _build_filters(self, entity: Entity, index: int) -> str:
-        prop_filters, prop_filter_params = parse_prop_clauses(
-            entity.properties, self._team.pk, prepend=str(index), allow_denormalized_props=True
-        )
-        self.params.update(prop_filter_params)
-        if entity.properties:
-            return prop_filters
-        return ""
