+++ ee/clickhouse/queries/funnels/funnel_trends_persons.py
@@ -1,0 +1,6 @@
+from ee.clickhouse.queries.funnels.base import ClickhouseFunnelBase
+
+
+class ClickhouseFunnelPersons(ClickhouseFunnelBase):
+    def get_query(self, format_properties):
+        pass
