+++ ee/clickhouse/queries/funnels/test/test_funnel.py
@@ -1,0 +1,26 @@
+from uuid import uuid4
+
+from ee.clickhouse.models.event import create_event
+from ee.clickhouse.queries.funnels.funnel import ClickhouseFunnel
+from ee.clickhouse.util import ClickhouseTestMixin
+from posthog.models.person import Person
+from posthog.queries.test.test_funnel import funnel_test_factory
+
+FORMAT_TIME = "%Y-%m-%d 00:00:00"
+MAX_STEP_COLUMN = 0
+COUNT_COLUMN = 1
+PERSON_ID_COLUMN = 2
+
+
+def _create_person(**kwargs):
+    person = Person.objects.create(**kwargs)
+    return Person(id=person.uuid, uuid=person.uuid)
+
+
+def _create_event(**kwargs):
+    kwargs.update({"event_uuid": uuid4()})
+    create_event(**kwargs)
+
+
+class TestFunnel(ClickhouseTestMixin, funnel_test_factory(ClickhouseFunnel, _create_event, _create_person)):  # type: ignore
+    pass
